moved routes. redis, and websocket to seperate files
